% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addTableToPlot.R
\name{addTableToPlot}
\alias{addTableToPlot}
\title{Add table to melting curve plots.}
\usage{
addTableToPlot(plotObj, tableDF, meltVar, clrs)
}
\arguments{
\item{plotObj}{ggplot object to which table should be added.}

\item{tableDF}{dataframe containing the table to be added.}

\item{meltVar}{character specifying the identifier column in tableDF.}

\item{clrs}{character vector of plot colors for each row in the table.}
}
\value{
ggplot object with table added to an area below the melting curves.
}
\description{
Adds a table of melting curve parameters to an existing
ggplot object containing the melting curve plots.
}
\details{
The table is added to an area below the melting curves.

\code{plotObj} is a ggplot object that contains the melting
curves of a protein estimated in different experiments.

 \code{tableDF} contains the curve parameters to be displayed in the
 table. The column containing the names of the experimental groups is
 specified by \code{meltVar}.
}
\seealso{
\code{\link{plotMeltingCurve}}
}

