% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/applyCoeffs.R
\name{applyCoeffs}
\alias{applyCoeffs}
\title{Normalize protein fold changes obtained by a high-throughput TPP-TR
experiment.}
\usage{
applyCoeffs(data, coeffs)
}
\arguments{
\item{data}{ExpressionSet containing matrix of protein fold changes for
normalization. Each row corresponds to a protein and each column to a
temperature condition.}

\item{coeffs}{Numeric vector or normalization coefficients for each temperature
condition. Length must correspond to number of columns in data.}
}
\value{
ExpressionSet containing the matrix of normalized fold changes.
}
\description{
Normalize fold changes stored in an ExpressionSet object using
a given vector of coefficients.
}
\details{
This function performs column-wise normalization of the protein fold changes in an expressionSet using a given set of normalization coefficients (also called 'correction factors').
Normalization is performed by multiplying each fold change with the corresponding coefficient.
}
\seealso{
\code{\link{tpptrNormalize}}
}

