% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checkCPUs.R
\name{checkCPUs}
\alias{checkCPUs}
\title{Check the user-specified number of CPUS.}
\usage{
checkCPUs(cpus)
}
\arguments{
\item{cpus}{Numeric value with desired maximum number of CPUS, or character
string with value 'max'.}
}
\value{
Maximal number of cores detected on the current device.
}
\description{
Assistant function to check the desired number of CPUs for parallel
melting curve fitting.
}
\details{
If the given number exceeds the available CPUS on the device, the maximum
available CPUs will be used instead. If the value 'max' was chosen (the default in
\code{\link{analyzeTPPTR}} and \code{\link{tpptrCurveFit}}), the maximum available
CPUS will be used as well.
}

