% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/computeNormFactors.R
\name{computeNormFactors}
\alias{computeNormFactors}
\title{Compute normalization coefficients for TTP-TR experiments}
\usage{
computeNormFactors(data, startPars, maxAttempts)
}
\arguments{
\item{data}{List of expressionSets with fold change measurements of normP.}

\item{startPars}{start values for the melting curve parameters. Will be passed
to function \code{\link{nls}} for curve fitting.}

\item{maxAttempts}{maximal number of curve fitting attempts if model does not converge.}
}
\value{
A list with nomalization curves for each replicate plus additional
information.
}
\description{
Fit sigmoids to median fold changes of each treatment group and
determine best fit.
}
\details{
If the melting curve fitting procedure does not converge, it will be
repeatedly started from perturbed starting parameters (maximum iterations
defined by argment \code{maxAttempts}).
}

