% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/computePvalues.R
\name{computePvalues}
\alias{computePvalues}
\title{Compare proteins of vehicle and treatment groups.}
\usage{
computePvalues(ids, minSlopes, mpDiffs, binWidth)
}
\arguments{
\item{ids}{character vector of protein ids.}

\item{minSlopes}{numeric vector of minimal slopes.}

\item{mpDiffs}{numeric vector of melting point differences.}

\item{binWidth}{bin width used for p-value computation.}
}
\description{
Compute p-values of melting point differences
}
\details{
Proteins are assigned to bins of specified size (default: 300)
according to their ranked minimal slopes. For each bin, p-values are computed
as described by Cox et al. (2008).
}
\references{
Cox, J., & Mann, M. (2008). MaxQuant enables high peptide
identification rates, individualized ppb-range mass accuracies and
proteome-wide protein quantification. Nature biotechnology, 26(12), 1367-1372.
}

