% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exprSubset.R
\name{exprSubset}
\alias{exprSubset}
\title{Select specified featureNames in an ExpressionSet}
\usage{
exprSubset(exprSet, subset)
}
\arguments{
\item{exprSet}{...}

\item{subset}{...
Creates a subset of an ExpressionSet that contains only specified featureNames.}
}
\description{
Select specified featureNames in an ExpressionSet
}

