% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fctSigmoidCCR.R
\name{fctSigmoidCCR}
\alias{fctSigmoidCCR}
\title{Sigmoidal function for dose response curve fitting}
\usage{
fctSigmoidCCR()
}
\value{
Character of the function.
}
\description{
Return sigmoidal function for curve fit in TPP-CCR experiments
}
\details{
the complete function has the following form:
y(bottom, top, infl, hill) <- bottom + (top - bottom) / (1 + exp((infl - x) * hill))
 Here, we set top=1 and bottom=0, which leaves two parameters for fitting:
y(infl, hill) <- 0 + (1 - 0) / (1 + exp((infl - x) * hill))
}

