% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fctSigmoidTR.R
\name{fctSigmoidTR}
\alias{fctSigmoidTR}
\title{Sigmoidal function for melting curve fitting}
\usage{
fctSigmoidTR(deriv)
}
\arguments{
\item{deriv}{derivative to return: 0 (default), 1, or 2.}
}
\value{
Function string.
}
\description{
Return sigmoidal function or its derivatives used for curve fit
in TPP-TR experiments
}

