% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filterFCs.R
\name{filterFCs}
\alias{filterFCs}
\title{Filter fold change columns for normalization set construction}
\usage{
filterFCs(data, pos, lb, ub)
}
\arguments{
\item{data}{expressionSet with fold changes for filtering}

\item{pos}{vector of FC-column positions for filtering}

\item{lb}{vector of lower bounds}

\item{ub}{vector of upper bounds}
}
\description{
Filter all data sets by specified criteria on fold change columns
}
\details{
...
}
\references{
...
}
\seealso{
...

Other ...: \code{\link{filterOther}};
  \code{\link{filterTables}}
}

