% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filterOther.R
\name{filterOther}
\alias{filterOther}
\title{Filter metadata columns for normalization set construction}
\usage{
filterOther(data, cols, lb, ub)
}
\arguments{
\item{data}{expressionSet with fold changes for filtering}

\item{cols}{vector of non-FC column names for filtering}

\item{lb}{vector of lower bounds}

\item{ub}{vector of upper bounds}
}
\description{
Filter all data sets by specified criteria on metadata columns
}
\details{
...
}
\references{
...
}
\seealso{
...

Other ...: \code{\link{filterFCs}};
  \code{\link{filterTables}}
}

