% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filterTables.R
\name{filterTables}
\alias{filterTables}
\title{Filter data for normalization set construction}
\usage{
filterTables(data, normReqs)
}
\arguments{
\item{data}{...}

\item{normReqs}{...}
}
\description{
Filter all data sets by specified criteria on fold changes and metadata columns
}
\details{
...
}
\references{
...
}
\seealso{
...

Other ...: \code{\link{filterFCs}};
  \code{\link{filterOther}}
}

