% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findJointP.R
\name{findJointP}
\alias{findJointP}
\title{Find commonly occurring proteins in each TPP experiment}
\usage{
findJointP(data)
}
\arguments{
\item{data}{list of expressionSets from which to find joint subset.}
}
\value{
A character vector of protein IDs in jointP.
}
\description{
Find proteins that commonly occur in all experiments.
}
\details{
\code{findJointP} is invoked in order to construct normalization set.
}

