% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fitMeltCurvesToProtein.R
\name{fitMeltCurvesToProtein}
\alias{fitMeltCurvesToProtein}
\title{Fit melting curves for each condition and replicate to a single protein}
\usage{
fitMeltCurvesToProtein(xMat, yDF, startPars, maxAttempts, expNames, resultPath,
  plotPathRel, protID, plotTheme, grConds, grReps, doPlot, addLegend)
}
\arguments{
\item{xMat}{temperature values for curve fit}

\item{yDF}{fold change values for curve fit}

\item{startPars}{start values for the melting curve parameters. Will be passed
to function \code{\link{nls}} for curve fitting.}

\item{maxAttempts}{maximal number of curve fitting attempts if model does not converge.}

\item{expNames}{necessary for construction of output dataframe}

\item{resultPath}{location where to store the melting curve plots.}

\item{plotPathRel}{relative plot locations}

\item{protID}{protein ids}

\item{plotTheme}{ggplot theme}

\item{grConds}{biological condition, can be used for paired colouring of melting curves}

\item{grReps}{experimental replicates}

\item{doPlot}{perform plotting?}

\item{addLegend}{boolean value indicating whether to add legend to the plot.}
}
\description{
...
}
\details{
\code{fitMeltCurvesToProtein} is invoked during parallelization.
Currently, adding figure legens (argument \code{addLegend}) is discouraged on
MAC OS-X because it can hamper parallelization.
}
\seealso{
\code{\link{tpptrCurveFit}}
}

