% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fitSigmoidCCR.R
\name{fitSigmoidCCR}
\alias{fitSigmoidCCR}
\title{Fit sigmoidal model to a vector of TPP-CCR measurements}
\usage{
fitSigmoidCCR(xVec, yVec, hill_init, pec50_init, slopeBounds, concBounds)
}
\arguments{
\item{xVec}{...}

\item{yVec}{...}

\item{hill_init}{...}

\item{pec50_init}{...}

\item{slopeBounds}{Bounds on the slope parameter for dose response curve fitting.}

\item{concBounds}{...}
}
\description{
...
}
\details{
...
}
\seealso{
\code{\link{fctSigmoidCCR}}, \code{\link{fctSigmoidTR}}, \code{\link{fitSigmoidTR}}
}

