% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fitSigmoidTR.R
\name{fitSigmoidTR}
\alias{fitSigmoidTR}
\title{Fit sigmoidal model to a vector of TPP-TR measurements}
\usage{
fitSigmoidTR(xVec, yVec, startPars, maxAttempts)
}
\arguments{
\item{xVec}{temperature values}

\item{yVec}{fold change values}

\item{startPars}{start values for the melting curve parameters. Will be passed
to function \code{\link{nls}} for curve fitting.}

\item{maxAttempts}{maximal number of curve fitting attempts if model does not converge.}
}
\value{
Fitted model.
}
\description{
Fit sigmoidal model to a vector of TPP-TR measurements
}

