% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/guessDRparamBoundaries.R
\name{guessDRparamBoundaries}
\alias{guessDRparamBoundaries}
\title{Compute concentration bounds for sigmoidal curve fitting}
\usage{
guessDRparamBoundaries(concentrations)
}
\arguments{
\item{concentrations}{Drug concentrations applied to aliquots in TPP-CCR experiment.}
}
\description{
Compute concentration bounds using largest dilution step among the concentration gradient.
}

