% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/guessInitialDRslope.R
\name{guessInitialDRslope}
\alias{guessInitialDRslope}
\title{Guess initial Hill slope}
\usage{
guessInitialDRslope(dose, response, hill_bds, cpd_effect)
}
\arguments{
\item{dose}{...}

\item{response}{...}

\item{hill_bds}{...}

\item{cpd_effect}{...}
}
\description{
Guess initial Hill slope for logistic curve fit of TPP-CCR experiments.
}
\details{
The estimated value will be used as starting parameter for the curve fitting procedure.
}

