% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/guessInitialpEC50.R
\name{guessInitialpEC50}
\alias{guessInitialpEC50}
\title{Guess initial pEC50 value}
\usage{
guessInitialpEC50(dose, response, conc_bds)
}
\arguments{
\item{dose}{...}

\item{response}{...}

\item{conc_bds}{...}
}
\description{
Guess initial pEC50 value for logistic curve fit of TPP-CCR experiments.
}
\details{
The estimated value will be used as starting parameter for the curve fitting procedure.
}

