% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hdacCCR1_smallExample.R
\docType{data}
\name{hdacCCR1_smallExample}
\alias{hdacCCR1_smallExample}
\title{Example subset of the Panobinostat TPP-CCR dataset (replicate 1)}
\description{
The first 500 entries of the dataset \link{hdacCCR1}.
}
\details{
You can use this dataset to explore the \code{\link{TPP}} package
functionalities without invoking the whole time consuming analysis on the big dataset.
}

