% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hdacTR_smallExample.R
\docType{data}
\name{hdacTR_smallExample}
\alias{hdacTR_smallExample}
\title{Example subset of the Panobinostat TPP-TR dataset}
\description{
The first 500 entries of the dataset \link{hdacTR}.
}
\details{
You can use this dataset to explore the \code{\link{TPP}} package
functionalities without invoking the whole time consuming analysis on the big dataset.
}

