% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/importCheckConditions.R
\name{importCheckConditions}
\alias{importCheckConditions}
\title{Assign generic condition information to each experiment, if not specified by the user.}
\usage{
importCheckConditions(condInfo, expectedLength)
}
\arguments{
\item{condInfo}{...}

\item{expectedLength}{...}
}
\description{
Assign generic condition information to each experiment, if not specified by the user.
}

