% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/importCheckConfigTable.R
\name{importCheckConfigTable}
\alias{importCheckConfigTable}
\title{Check whether obligatory experiment information is provided via data
frame or spreadsheet file.}
\usage{
importCheckConfigTable(infoTable)
}
\arguments{
\item{infoTable}{dataframe, or character object with the path to a
file, that specifies important details of the TPP-CCR experiment. See
Section \code{details} for instructions how to create this object.}
}
\description{
Check whether obligatory experiment information is provided via data
frame or spreadsheet file.
}

