% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/importCheckDataFormat.R
\name{importCheckDataFormat}
\alias{importCheckDataFormat}
\title{Check whether dataframes of filenames are specified for data import}
\usage{
importCheckDataFormat(files, dataframes, expNames, genericExpNames)
}
\arguments{
\item{files}{...}

\item{dataframes}{...}

\item{expNames}{...}

\item{genericExpNames}{...}
}
\description{
Check whether dataframes of filenames are specified for data import
}

