% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/importCheckExperimentNames.R
\name{importCheckExperimentNames}
\alias{importCheckExperimentNames}
\title{Assign generic experiment names.}
\usage{
importCheckExperimentNames(expNames, expectedLength)
}
\arguments{
\item{expNames}{...}

\item{expectedLength}{...}
}
\description{
Assign generic experiment names.
}

