% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/importCheckLabels.R
\name{importCheckLabels}
\alias{importCheckLabels}
\title{Check label argument for the import function}
\usage{
importCheckLabels(labelValues, temperatureMat)
}
\arguments{
\item{labelValues}{...}

\item{temperatureMat}{...}
}
\description{
Check label argument for the import function
}

