% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/importCheckReplicates.R
\name{importCheckReplicates}
\alias{importCheckReplicates}
\title{Assign generic replicate information to each experiment, if not specified by the user.}
\usage{
importCheckReplicates(repInfo, expectedLength)
}
\arguments{
\item{repInfo}{...}

\item{expectedLength}{...}
}
\description{
Assign generic replicate information to each experiment, if not specified by the user.
}

