% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/importCheckTemperatures.R
\name{importCheckTemperatures}
\alias{importCheckTemperatures}
\title{Determine matrix of temperatures corresponding to the TMT labels}
\usage{
importCheckTemperatures(temp, nData)
}
\arguments{
\item{temp}{...}

\item{nData}{...}
}
\description{
Determine matrix of temperatures corresponding to the TMT labels
}

