% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/importSingleExp.R
\name{importSingleExp}
\alias{importSingleExp}
\title{Import TPP-TR or TPP-CCR table}
\usage{
importSingleExp(dataframe = NULL, filename = NULL, labels, labelValues,
  name, condition, replicate, idVar, fcStr, qualColName, naStrs, type)
}
\arguments{
\item{dataframe}{...}

\item{filename}{...}

\item{labels}{...}

\item{labelValues}{numeric vector of values that distinguish each isotope
label. Contains temperature conditions when importing a TPP-TR experiment, and
concentrations when importing an TPP-CCR experiment.}

\item{name}{character string with name of the imported experiment.}

\item{condition}{character string naming the experimental condition belonging
to an imported TPP-TR dataset.}

\item{replicate}{character string naming the experimental replicate belonging
to an imported TPP-TR dataset.}

\item{idVar}{character string indicating which data column provides the unique
identifiers for each protein.}

\item{fcStr}{character string indicating which columns contain the actual
fold change values. Those column names containing the suffix
\code{fcStr} will be regarded as containing fold change values.}

\item{qualColName}{character string indicating which column can be used for
additional quality criteria when deciding between different non-unique
protein identifiers.}

\item{naStrs}{character vector indicating missing values in the data table.
When reading data from file, this value will be passed on to the argument
\code{na.strings}
in function \code{read.delim}.}

\item{type}{character describing the type of experiment to be imported. Can
either be "CCR" or "TR". The output expressionSet object will be adjusted
accordingly.}
}
\description{
Import a single TPP-CR experiment from file or data frame and convert
it into an ExpressionSet.
}
\details{
Called by functions \code{\link{tpptrImport}} and \code{\link{tppccrImport}}.
}
\seealso{
\code{\link{tpptrImport}}, \code{\link{tppccrImport}}.
}

