% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inflectionPoint.R
\name{inflectionPoint}
\alias{inflectionPoint}
\title{Compute inflection point of sigoidal model}
\usage{
inflectionPoint(model, xRange)
}
\arguments{
\item{model}{...}

\item{xRange}{...}
}
\description{
Compute inflection point of sigoidal model
}

