% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meltCurveParamNames.R
\name{meltCurveParamNames}
\alias{meltCurveParamNames}
\title{Assistant function that returns the column names of the melting curve parameters
in the internal datasets.}
\usage{
meltCurveParamNames(returnParNames = T, returnPerformanceInfo = T)
}
\arguments{
\item{returnParNames}{...}

\item{returnPerformanceInfo}{...}
}
\description{
Assistant function that returns the column names of the melting curve parameters
in the internal datasets.
}
\details{
This function is intended to assure consistency when accessing, manipulating, or
storing melting curve paramter columns in the package's data objects.
}

