% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meltingCurveSlope.R
\name{meltingCurveSlope}
\alias{meltingCurveSlope}
\title{Compute melting point slope of sigmoidal model}
\usage{
meltingCurveSlope(model, xInfl)
}
\arguments{
\item{model}{...}

\item{xInfl}{...}
}
\description{
Compute melting point slope of sigmoidal model
}

