% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meltingPoint.R
\name{meltingPoint}
\alias{meltingPoint}
\title{Compute melting point of sigmoidal model}
\usage{
meltingPoint(model, xRange)
}
\arguments{
\item{model}{model of fitted melting curve}

\item{xRange}{temperature range in which to search}
}
\description{
Compute melting point of sigmoidal model
}

