% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mpSinglePval.R
\name{mpSinglePval}
\alias{mpSinglePval}
\title{Compute p-value for melting point difference value}
\usage{
mpSinglePval(x, r_1, r0, r1)
}
\arguments{
\item{x}{...}

\item{r_1}{...}

\item{r0}{...}

\item{r1}{...}
}
\description{
Compute p-value for a given melting point difference value using
precomputed distribution quantiles.
}
\references{
Cox, J., & Mann, M. (2008). MaxQuant enables high peptide
identification rates, individualized ppb-range mass accuracies and
proteome-wide protein quantification. Nature biotechnology, 26(12), 1367-1372.
}

