% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/paramsSigmoid.R
\name{paramsSigmoid}
\alias{paramsSigmoid}
\title{Compute melting curve and fit parameters for sigmoidal model}
\usage{
paramsSigmoid(model, xRange, y)
}
\arguments{
\item{model}{...}

\item{xRange}{...}

\item{y}{...}
}
\description{
Compute melting curve and fit parameters for sigmoidal model
}

