% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotColors.R
\name{plotColors}
\alias{plotColors}
\title{Create matching color pairs for vehicle and treatment groups}
\usage{
plotColors(expConditions, expReplicates)
}
\arguments{
\item{expConditions}{...}

\item{expReplicates}{...}
}
\description{
Create matching color pairs for vehicle and treatment groups
}

