% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotMeltingCurve.R
\name{plotMeltingCurve}
\alias{plotMeltingCurve}
\title{Plot melting curve fitted to TPP-TR fold changes}
\usage{
plotMeltingCurve(modelList, xMat, fcMat, curvePars, protID, filename, plotTheme,
  expConditions, expReplicates, addLegend)
}
\arguments{
\item{modelList}{list of fitted melting curve models. Field names must
correspond to experiment names.}

\item{xMat}{matrix of temperature values (row-wise) for each experiment.
Rownames must correspond to experiment names.}

\item{fcMat}{matrix of fold change values (row-wise) for each experiment.
Rownames must correspond to experiment names.}

\item{curvePars}{matrix of estimated melting curve parameters (row-wise) for
each experiment. Rownames must correspond to experiment names.}

\item{protID}{name of current protein. Will appear in title plot.}

\item{filename}{file name of plot. If NULL, plot will be discarded.}

\item{plotTheme}{ggplot theme for current plot}

\item{expConditions}{experiment conditions (typically
"Treatment" or "Vehicle")}

\item{expReplicates}{experimental replicates (typically 1,2, ...)}

\item{addLegend}{Add legend to plots. Legend is disabled if OS is not Windows or Linux.}
}
\description{
Plot melting curve fitted to TPP-TR fold changes
}
\details{
Produce a ggplot to vizualise fitted melting curves, melting points and slopes.
}

