% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotMinSlopes_vs_MPDiffs.R
\name{plotMinSlopes_vs_MPDiffs}
\alias{plotMinSlopes_vs_MPDiffs}
\title{Create 'bottle plot' and histogram of minimal slopes}
\usage{
plotMinSlopes_vs_MPDiffs(mpDiffs, minSlopes, pValues, expName1, expName2,
  plotTheme = tppDefaultTheme(), path = NULL)
}
\arguments{
\item{mpDiffs}{numeric vector of melting point differences.}

\item{minSlopes}{numeric vector of minimal slopes.}

\item{pValues}{numeric vector of p-values}

\item{expName1}{name of the first experiment for the plot title.}

\item{expName2}{name of the second experiment for the plot title.}

\item{plotTheme}{ggplot theme}

\item{path}{destination where to save the plot (optional)}
}
\description{
Create 'bottle plot' and histogram of minimal slopes
}

