% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotNormCurves.R
\name{plotNormCurves}
\alias{plotNormCurves}
\title{Plot QC plot to visualize curve fit of normalization curves to fold change medians.}
\usage{
plotNormCurves(modelList, xMat, fcMat, r2Vec, nNormP, plotTheme, pathPlot)
}
\arguments{
\item{modelList}{...}

\item{xMat}{...}

\item{fcMat}{...}

\item{r2Vec}{...}

\item{nNormP}{...}

\item{plotTheme}{...}

\item{pathPlot}{...}
}
\description{
Plot QC plot to visualize curve fit of normalization curves to fold change medians.
}

