% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pValsCurrentBin.R
\name{pValsCurrentBin}
\alias{pValsCurrentBin}
\title{Determine meltpoints outside of 0.16 and 0.86 percentiles for each bin}
\usage{
pValsCurrentBin(mpDiffs)
}
\arguments{
\item{mpDiffs}{numeric vector of melting point differences.}
}
\description{
Determine meltpoints outside of 0.16 and 0.86 percentiles for each bin
}
\references{
Cox, J., & Mann, M. (2008). MaxQuant enables high peptide
identification rates, individualized ppb-range mass accuracies and
proteome-wide protein quantification. Nature biotechnology, 26(12), 1367-1372.
}

