% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/removeDuplicates.R
\name{removeDuplicates}
\alias{removeDuplicates}
\title{Remove duplicate entries in ID column}
\usage{
removeDuplicates(inDF, refColName, nonNAColNames, qualColName = NULL)
}
\arguments{
\item{inDF}{NA}

\item{refColName}{NA}

\item{nonNAColNames}{NA}

\item{qualColName}{character string indicating which column can be used for
additional quality criteria when deciding between different non-unique
protein identifiers.}
}
\description{
Remove duplicate entries in ID column
}

