% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/resultFilterCurvePars.R
\name{resultFilterCurvePars}
\alias{resultFilterCurvePars}
\title{Check melting curve quality before starting p-value computation.}
\usage{
resultFilterCurvePars(r2V, r2T, plV, minR2 = 0.8, maxPlateau = 0.3)
}
\arguments{
\item{r2V}{...}

\item{r2T}{...}

\item{plV}{...}

\item{minR2}{...}

\item{maxPlateau}{...}
}
\description{
Check melting curve quality before starting p-value computation.
}

