% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/robustNlsPredict.R
\name{robustNlsPredict}
\alias{robustNlsPredict}
\title{Robust nls prediction that can handle non-converged models.}
\usage{
robustNlsPredict(model, newdata = NULL)
}
\arguments{
\item{model}{...}

\item{newdata}{...}
}
\description{
A wrapper function for predict.nls() that checks whether model actually converged before starting the prediction.
}

