% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/storeMeltCurveParams.R
\name{storeMeltCurveParams}
\alias{storeMeltCurveParams}
\title{Store estimated melting curve parameters in featureData}
\usage{
storeMeltCurveParams(data, params)
}
\arguments{
\item{data}{expressionSet with TPP-TR experiment data.}

\item{params}{dataframe of estimated melting curve parameters.}
}
\description{
Store estimated melting curve parameters in featureData
}

