% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tppccrNormalize.R
\name{tppccrNormalize}
\alias{tppccrNormalize}
\title{Normalize data from TPP-CCR experiments}
\usage{
tppccrNormalize(data)
}
\arguments{
\item{data}{expressionSet with measurements to be normalized}
}
\description{
Normalize each fold change column by its median.
}

