% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tppccrTransform.R
\name{tppccrTransform}
\alias{tppccrTransform}
\title{Transform fold changes of TPP-CCR experiment}
\usage{
tppccrTransform(data, fcCutoff = 1.5)
}
\arguments{
\item{data}{expressionSet object containing the data to be transformed.}

\item{fcCutoff}{Cutoff for highest compound concentration fold change.}
}
\description{
Transform fold changes of TPP-CCR experiment
}

