% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tppExport.R
\name{tppExport}
\alias{tppExport}
\title{Produce Excel table of TPP-TR or TPP-CCR experiment.}
\usage{
tppExport(tab, file)
}
\arguments{
\item{tab}{Table with results of the TPP analysis.}

\item{file}{path for storing results table}
}
\description{
Produce Excel table of TPP-TR or TPP-CCR experiment.
}

