% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tppQCPlotsCorrelateExperiments.R
\name{tppQCPlotsCorrelateExperiments}
\alias{tppQCPlotsCorrelateExperiments}
\title{Plot pairwise relationships between the proteins in different TPP experiments.}
\usage{
tppQCPlotsCorrelateExperiments(tppData, annotStr = "",
  ggplotTheme = tppDefaultTheme(), writePdf = TRUE, path = getwd())
}
\arguments{
\item{tppData}{List of expressionSets with data to be plotted.}

\item{annotStr}{String with additional information to be added to the plot.}

\item{ggplotTheme}{ggplot theme for the created plots.}

\item{writePdf}{Boolean value. Should plots be saved to file?}

\item{path}{Location where to store resulting plot.}
}
\description{
Plot pairwise relationships between the proteins in different TPP experiments.
}
\seealso{
\code{\link{tppDefaultTheme}}
}

