% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tpptrQCPlotsMinSlopes_vs_MPdiffs.R
\name{tpptrQCPlotsMinSlopes_vs_MPdiffs}
\alias{tpptrQCPlotsMinSlopes_vs_MPdiffs}
\title{Plot minSlope vs. melting point differences}
\usage{
tpptrQCPlotsMinSlopes_vs_MPdiffs(resultTable, expNames = NULL,
  expRepl = NULL, expCond = NULL)
}
\arguments{
\item{resultTable}{Table of the TPP-TR results.}

\item{expNames}{Names of each experiment.}

\item{expRepl}{Replicate each experiment belongs to.}

\item{expCond}{Condition (Treatment or Vehicle) each experiment belongs to.}
}
\description{
\code{tpptrQCPlotsMinSlopes_vs_MPdiffs} plots the minSlope vs. melting
point difference for each protein and highlights proteins with significant shifts.
}

